+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.0.0-alpha.10] - 2025-08-21
+### Changed
+- Added support for **view files** 
+  - Added `/:siteId/:pageUrl/view-file/:elementName` route (`get`)
+  - Added links on fileView both on client side and server side (opening in new tab)
+  - Added links on `review` page to view files (opening in new tab)
+  - Added unit tests
+- Changed delete file url to `/:siteId/:pageUrl/delete-file/:elementName`
+
 ## [v1.0.0-alpha.9] - 2025-08-19
 ### Changed
 - Added support for **delete files**